+++ src/pages/IntakeForms/IntakeForm/index.tsx
@@ -1,13 +1,13 @@
-import React from "react";
+import { useRouter } from "next/router";
+import React, { useState } from "react";
 import IntakeForm from "../../../components/IntakeForm/IntakeForm";
 
-export default function Citizenship() {
+export default function IntakeScreen() {
+  const router = useRouter();
   return (
     <div>
       <div>
-        <IntakeForm
-        // title="Citizenship"
-        />
+        <IntakeForm/>
       </div>
     </div>
   );
